Pugs Parents logo
Introducing the Pugs NFT Ecosystem on TON, a unique collection where every Pug lives through a continuous life cycle. It starts with the pre sale for Pugs Puppies which we called Pugs Parents, we did not want to leave you empty handed or with a placeholder NFT, so we let a renowned TON artist create Pug Parents. After a succesfull pre sale we will proceed with Pugs Puppies, designed by the talented Ilya Polutis.